Improvements to UI prt 1.
Added new assets to use in UI. They are used in "Quiz" section as an information box replacement and answer sheet holder.
Replaced and reworked "Timeline" and "Quiz" buttons to fit the image of the UI designer--somewhat. All buttons should still be functional, had they were in the first place.
Inside the code for the quiz buttons, I added general names so that it can be reused if needed.
Edited font and font size throughout the two scenes.
Placed sample text for testing purposes.